Add video functionality
To add a new video, we use the Youtube API, with this, we can made a get
equest and get the title of the video, the id and the description, with
this way we just have to ask for the video id and rails make the rest.

Also, we ask for a category of the video, because a video has one
category, and it is necessary specfy it. The two fields are validates
(cannot be empty, or if the video already exist).

But to do this possible, we implemented the Faraday gem, this gem
permit create HTTP requests.